+++ scripts/codegen/cldr-parser.go
@@ -10,14 +10,24 @@ import (
 func parseCldrData(languageLocalesMap map[string][]string) error {
 	for language, locales := range languageLocalesMap {
 		log.Info().Msgf("parsing data for language %s", language)
+
 		_, err := parseCldrGregorianData(language)
 		if err != nil {
 			return err
 		}
 
+		_, err = parseCldrDateFieldsData(language)
+		if err != nil {
+			return err
+		}
+
 		for _, locale := range locales {
-			log.Info().Msgf("parsing data for locale %s", locale)
-			_, err := parseCldrGregorianData(locale)
+			_, err = parseCldrGregorianData(locale)
+			if err != nil {
+				return err
+			}
+
+			_, err = parseCldrDateFieldsData(locale)
 			if err != nil {
 				return err
 			}
@@ -36,7 +46,7 @@ func parseCldrGregorianData(locale string) (*CldrGregorianData, error) {
 	f, err := os.Open(fPath)
 	if err != nil {
 		if os.IsNotExist(err) {
-			log.Warn().Msgf("cldr data for locale %s not exist, skipped", locale)
+			log.Warn().Msgf("gregorian data for locale %s not exist, skipped", locale)
 			return nil, nil
 		}
 		return nil, err
@@ -60,6 +70,32 @@ func parseCldrGregorianData(locale string) (*CldrGregorianData, error) {
 		if value, ok := v["_value"].(string); ok {
 			mainData.Dates.Calendars.Gregorian.DateFormats.Short = strings.ToUpper(value)
 		}
+	}
+
+	return &data, nil
+}
+
+func parseCldrDateFieldsData(locale string) (*CldrDateFieldsData, error) {
+	// Prepare file path
+	cldrDatesFullDir := filepath.Join(RAW_DIR, "cldr-dates-full", "main")
+	fPath := filepath.Join(cldrDatesFullDir, locale, "dateFields.json")
+
+	// Open file
+	f, err := os.Open(fPath)
+	if err != nil {
+		if os.IsNotExist(err) {
+			log.Warn().Msgf("date fields data for locale %s not exist, skipped", locale)
+			return nil, nil
+		}
+		return nil, err
+	}
+	defer f.Close()
+
+	// Parse JSON
+	var data CldrDateFieldsData
+	err = json.NewDecoder(f).Decode(&data)
+	if err != nil {
+		return nil, err
 	}
 
 	return &data, nil